manifest: better handling of path target
When passing a `.` path while inside a package directory, it passed to
path restrict generator as a simple path, which resulted in broken
restrict which collected all ebuilds in the repository.

By using `os.path.relpath` to base of repository, we make sure the
correct path is passed and the correct restricts are generated.

Fixes: https://github.com/pkgcore/pkgdev/issues/85
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>