bugs: support piping package list from stdin
When calling the tool without packages (for example just `pkgdev bugs`),
and you pipe (from a file or another command) a list of packages, it
will read the package list and use it.

The input file can have empty lines, and comments. Everything after the
first # is considered a comment. Whitespaces are stripped.

Resolves: https://github.com/pkgcore/pkgdev/issues/136
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>