bugs: improve best match finder
Use ``find_best_match`` for finding the best match, for the targets list
(instead of just selecting the max for each target). Apply this for
stabilization groups and cmd line packages. We do need to opt out the
preference for semi-stable over non-stable, as we want to stabilize
those requested.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>