tatt: fix required_use for packages from bug
When invoked on bug's package list, we use nattka to find matches in the
repo. Nattka requires configured repo, which we were passing beforehand.
But this results in a case where we have a configured packages, without
source ebuild's "required_use" (it was "configured" by the domain).

While somewhat ugly, fix it by performing a second iteration, this time
over the source repo, and get the expected package object we can use.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>